update_PredictAndApp

+++ prediction.py
@@ -15,10 +15,10 @@ def pred_class(model: torch.nn.Module, image: Image.Image, class_names: List[str
     transformed_image = image_transform(image).unsqueeze(dim=0)
 
     # Set the model to evaluation mode
-    model.eval()
+    model.eval().half()
     with torch.no_grad():
         # Make a prediction
-        target_image_pred = model(transformed_image)
+        target_image_pred = model(transformed_image).half()
         target_image_pred_probs = torch.softmax(target_image_pred, dim=1)
         target_image_pred_label = torch.argmax(target_image_pred_probs, dim=1)
